+++ lxd/migrate.go
@@ -309,9 +309,9 @@ func (s *migrationSourceWs) Do(migrateOp *operation) error {
 			idmap := IDMapType{
 				Isuid:    proto.Bool(ctnIdmap.Isuid),
 				Isgid:    proto.Bool(ctnIdmap.Isgid),
-				Hostid:   proto.Int(ctnIdmap.Hostid),
-				Nsid:     proto.Int(ctnIdmap.Nsid),
-				Maprange: proto.Int(ctnIdmap.Maprange),
+				Hostid:   proto.Int32(int32(ctnIdmap.Hostid)),
+				Nsid:     proto.Int32(int32(ctnIdmap.Nsid)),
+				Maprange: proto.Int32(int32(ctnIdmap.Maprange)),
 			}
 
 			idmaps = append(idmaps, &idmap)
@@ -647,9 +647,9 @@ func (c *migrationSink) Do(migrateOp *operation) error {
 			e := shared.IdmapEntry{
 				Isuid:    *idmap.Isuid,
 				Isgid:    *idmap.Isgid,
-				Nsid:     int(*idmap.Nsid),
-				Hostid:   int(*idmap.Hostid),
-				Maprange: int(*idmap.Maprange)}
+				Nsid:     int64(*idmap.Nsid),
+				Hostid:   int64(*idmap.Hostid),
+				Maprange: int64(*idmap.Maprange)}
 			srcIdmap.Idmap = shared.Extend(srcIdmap.Idmap, e)
 		}
